# Conflicts:
#	WebCrawler/src/main/java/com/crawler/controller/SpecialAlertController.java

+++ WebCrawler/src/main/java/com/crawler/controller/SpecialAlertController.java
@@ -8,7 +8,6 @@ import com.crawler.entity.dto.special_alert.SpecialAlertCreateDto;
 import com.crawler.entity.dto.special_alert.SpecialAlertEditDto;
 import com.crawler.entity.dto.special_alert.SpecialAlertInfoDto;
 import com.crawler.entity.dto.special_alert.SpecialAlertPageQueryDto;
-import com.crawler.mapper.UserMapper;
 import com.crawler.service.SpecialAlertService;
 import com.crawler.service.UserService;
 import jakarta.annotation.Resource;
@@ -49,7 +48,7 @@ public class SpecialAlertController {
     @GetMapping("/searchById")
     @RequirePermission(module = "alert", action = "alert_update")
     public Result searchById(HttpServletRequest request,
-                          @RequestParam Integer alertId){
+                          @RequestParam Long alertId){
 
         SpecialAlertSetting specialAlertInfo = specialAlertService.getSpecialAlertById(alertId);
 
@@ -63,7 +62,7 @@ public class SpecialAlertController {
     public Result create(HttpServletRequest request,
                          @RequestBody SpecialAlertCreateDto createDto) {
         User currentUser = (User) request.getAttribute("currentUser");
-        createDto.setUserId(Long.valueOf(currentUser.getUserId()));
+        createDto.setUserId(currentUser.getUserId());
 
         Map<String, Object> data = specialAlertService.create(createDto);
         return Result.success(data);
@@ -80,7 +79,7 @@ public class SpecialAlertController {
     //  启用 / 关闭预警专题（异步通知Python）
     @GetMapping("/triggerState")
     @RequirePermission(module = "alert", action = "alert_update")
-    public Result triggerState(@RequestParam Integer alertId) {
+    public Result triggerState(@RequestParam Long alertId) {
         Map<String, Object> data = specialAlertService.toggleTriggerState(alertId);
         // 返回最新状态码
         return Result.success(data);
@@ -89,7 +88,7 @@ public class SpecialAlertController {
     //  删除预警专题（专题须处于关闭状态）
     @RequirePermission(module = "alert", action = "alert_delete")
     @GetMapping("/alertDelete")
-    public Result alertDelete(@RequestParam Integer alertId) {
+    public Result alertDelete(@RequestParam Long alertId) {
         return specialAlertService.delete(alertId);
     }
