+++ WebCrawler/src/main/java/com/crawler/entity/dto/special_alert/SpecialAlertDto.java
@@ -11,7 +11,7 @@ import java.util.Date;
 @AllArgsConstructor
 @NoArgsConstructor
 public class SpecialAlertDto {
-    private Integer alertId;      // 专题ID（主键）
+    private Long alertId;      // 专题ID（主键）
     private Long userId;            // 用户ID（外键）
     private String alertName;     // 专题名称
     private Integer triggerState;   // 启用状态，0：已停止，1：已启用
